+++ flow/e2e/eventhub/eventhub_helper.go
@@ -1,4 +1,4 @@
-package e2e
+package e2e_eventhub
 
 import (
 	"context"
