+++ flow/e2e/sqlserver/qrep_flow_sqlserver_test.go
@@ -1,23 +1,92 @@
-package e2e
+package e2e_sqlserver
 
 import (
 	"context"
 	"fmt"
+	"os"
+	"testing"
 	"time"
 
+	"github.com/PeerDB-io/peer-flow/e2e"
 	"github.com/PeerDB-io/peer-flow/generated/protos"
 	"github.com/PeerDB-io/peer-flow/model"
 	"github.com/PeerDB-io/peer-flow/model/qvalue"
+	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/joho/godotenv"
+	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
+	"go.temporal.io/sdk/testsuite"
 )
 
-func (s *E2EPeerFlowTestSuite) setupSQLServerTable(tableName string) {
+const sqlserverSuffix = "s3"
+
+type PeerFlowE2ETestSuiteSQLServer struct {
+	suite.Suite
+	testsuite.WorkflowTestSuite
+
+	pool       *pgxpool.Pool
+	sqlsHelper *SQLServerHelper
+}
+
+func TestPeerFlowE2ETestSuiteSQLServer(t *testing.T) {
+	suite.Run(t, new(PeerFlowE2ETestSuiteSQLServer))
+}
+
+// setup sql server connection
+func (s *PeerFlowE2ETestSuiteSQLServer) setupSQLServer() {
+	env := os.Getenv("ENABLE_SQLSERVER_TESTS")
+	if env != "true" {
+		s.sqlsHelper = nil
+		return
+	}
+
+	sqlsHelper, err := NewSQLServerHelper("test_sqlserver_peer")
+	require.NoError(s.T(), err)
+	s.sqlsHelper = sqlsHelper
+}
+
+func (s *PeerFlowE2ETestSuiteSQLServer) SetupSuite() {
+	err := godotenv.Load()
+	if err != nil {
+		// it's okay if the .env file is not present
+		// we will use the default values
+		log.Infof("Unable to load .env file, using default values from env")
+	}
+
+	log.SetReportCaller(true)
+
+	pool, err := e2e.SetupPostgres(sqlserverSuffix)
+	if err != nil {
+		s.Fail("failed to setup postgres", err)
+	}
+	s.pool = pool
+
+	s.setupSQLServer()
+}
+
+// Implement TearDownAllSuite interface to tear down the test suite
+func (s *PeerFlowE2ETestSuiteSQLServer) TearDownSuite() {
+	err := e2e.TearDownPostgres(s.pool, sqlserverSuffix)
+	if err != nil {
+		s.Fail("failed to drop Postgres schema", err)
+	}
+
+	if s.sqlsHelper != nil {
+		err = s.sqlsHelper.CleanUp()
+		if err != nil {
+			s.Fail("failed to clean up sqlserver", err)
+		}
+	}
+}
+
+func (s *PeerFlowE2ETestSuiteSQLServer) setupSQLServerTable(tableName string) {
 	schema := getSimpleTableSchema()
 	err := s.sqlsHelper.CreateTable(schema, tableName)
 	require.NoError(s.T(), err)
 }
 
-func (s *E2EPeerFlowTestSuite) insertRowsIntoSQLServerTable(tableName string, numRows int) {
+func (s *PeerFlowE2ETestSuiteSQLServer) insertRowsIntoSQLServerTable(tableName string, numRows int) {
 	schemaQualified := fmt.Sprintf("%s.%s", s.sqlsHelper.SchemaName, tableName)
 	for i := 0; i < numRows; i++ {
 		params := make(map[string]interface{})
@@ -37,16 +106,15 @@ func (s *E2EPeerFlowTestSuite) insertRowsIntoSQLServerTable(tableName string, nu
 	}
 }
 
-func (s *E2EPeerFlowTestSuite) setupPGDestinationTable(schemaName, tableName string) {
+func (s *PeerFlowE2ETestSuiteSQLServer) setupPGDestinationTable(tableName string) {
 	ctx := context.Background()
-	_, err := s.pool.Exec(ctx, fmt.Sprintf("CREATE SCHEMA IF NOT EXISTS %s", schemaName))
+
+	_, err := s.pool.Exec(ctx, fmt.Sprintf("DROP TABLE IF EXISTS e2e_test_%s.%s", sqlserverSuffix, tableName))
 	require.NoError(s.T(), err)
 
-	_, err = s.pool.Exec(ctx, fmt.Sprintf("DROP TABLE IF EXISTS %s.%s", schemaName, tableName))
-	require.NoError(s.T(), err)
-
-	//nolint:lll
-	_, err = s.pool.Exec(ctx, fmt.Sprintf("CREATE TABLE %s.%s (id TEXT, card_id TEXT, v_from TIMESTAMP, price NUMERIC, status INT)", schemaName, tableName))
+	_, err = s.pool.Exec(ctx,
+		fmt.Sprintf("CREATE TABLE e2e_test_%s.%s (id TEXT, card_id TEXT, v_from TIMESTAMP, price NUMERIC, status INT)",
+			sqlserverSuffix, tableName))
 	require.NoError(s.T(), err)
 }
 
@@ -62,13 +130,13 @@ func getSimpleTableSchema() *model.QRecordSchema {
 	}
 }
 
-func (s *E2EPeerFlowTestSuite) Test_Complete_QRep_Flow_SqlServer_Append() {
+func (s *PeerFlowE2ETestSuiteSQLServer) Test_Complete_QRep_Flow_SqlServer_Append() {
 	if s.sqlsHelper == nil {
 		s.T().Skip("Skipping SQL Server test")
 	}
 
 	env := s.NewTestWorkflowEnvironment()
-	registerWorkflowsAndActivities(env)
+	e2e.RegisterWorkflowsAndActivities(env)
 
 	numRows := 10
 	tblName := "test_qrep_flow_avro_ss_append"
@@ -77,13 +145,13 @@ func (s *E2EPeerFlowTestSuite) Test_Complete_QRep_Flow_SqlServer_Append() {
 	s.setupSQLServerTable(tblName)
 	s.insertRowsIntoSQLServerTable(tblName, numRows)
 
-	s.setupPGDestinationTable(s.sqlsHelper.SchemaName, tblName)
-	dstTableName := fmt.Sprintf("%s.%s", s.sqlsHelper.SchemaName, tblName)
+	s.setupPGDestinationTable(tblName)
+	dstTableName := fmt.Sprintf("e2e_test_%s.%s", sqlserverSuffix, tblName)
 
-	//nolint:lll
-	query := fmt.Sprintf("SELECT * FROM %s.%s WHERE v_from BETWEEN {{.start}} AND {{.end}}", s.sqlsHelper.SchemaName, tblName)
+	query := fmt.Sprintf("SELECT * FROM %s.%s WHERE v_from BETWEEN {{.start}} AND {{.end}}",
+		s.sqlsHelper.SchemaName, tblName)
 
-	postgresPeer := GeneratePostgresPeer(postgresPort)
+	postgresPeer := e2e.GeneratePostgresPeer(e2e.PostgresPort)
 
 	qrepConfig := &protos.QRepConfig{
 		FlowJobName:                tblName,
@@ -100,7 +168,7 @@ func (s *E2EPeerFlowTestSuite) Test_Complete_QRep_Flow_SqlServer_Append() {
 		WaitBetweenBatchesSeconds:  5,
 	}
 
-	runQrepFlowWorkflow(env, qrepConfig)
+	e2e.RunQrepFlowWorkflow(env, qrepConfig)
 
 	// Verify workflow completes without error
 	s.True(env.IsWorkflowCompleted())
