+++ flow/e2e/sqlserver/sqlserver_helper.go
@@ -1,4 +1,4 @@
-package e2e
+package e2e_sqlserver
 
 import (
 	"context"
